Moved the chat boxes to be horizontal and use angular material to have them be able to  be expanded, almost like chat boxes.
Changed in server.js the rooms to be string concantnation so that whenever a new message is sent, it uses room+ groupNumber to create the group and emit messages within the group.

From last commit:
Displaying works through using a chatGroup.model.ts which has joined as a boolean to check
 if can be displayed for the client. has messages which is automatically updated within the subscription of the
chat.component.ts that was initilizaed within ngInit. also has the group number which is used to send
 out that number for  server.js to know which room to emit that message