Created database structure: Summary:
Users:
-user info
-Chats

Chats:
-messages: array of { creator, content, time}
-Users

No message object/collection because that would mean that all the messages would be stored in one place, which would require you to do Message.find() which will be very slow considering there will be so many messages.

In essence, Chat and User are normalized (ie uses Chat _id or User _id) so that we can query between the objects which are both quite large.

Chat's messages are denormalized (ie immediately stored into the database document) because of reasoning above.

To do:
continue creating the database fetch methods, creating socket rooms for one to one chat, bringing it all to the front end.