psts1. in userList html component, person clicks on user, sending the userId to directly to profile component... user's id: localhost3000/profile/:userId348043, renavigates directly to that path. psts1b. in chat clicking on user name in chat box     psts1c) in ngonint parammap gets the parameter userId34093455 psts2. search in main service for userId348043: returns it profile component to display done within ngoninit
Early stages of profile. just to view the basics of the user's information.

in future.
make get requests for the profile because currently the program gets all users when logging in but
optimally should just be getting friends. other users that you want to talk to should be done through a get request to server.